Correct find_min() to check array size.
* Find_min was checking size (the function), not size(), and getting an
IndexError, so the test passed, but not for the reasons I assumed;
it was because top was set to limit + 1 so it was always out of bounds.
* Change top to be limit (and not limit + 1) by default.
* Find_min should not be checking size() either, because that isn’t related
to the number of items in the array, just the capacity. Instead, make a new
internal variable called item_count, and make find_min check this value
instead of size().