Changes made per code review.
Change ‘key’ to ‘priority’ and ‘value’ to ‘data’.
Change node’s ‘child’ to ‘next_item’.
Stop converting float indices to ints; raise an error instead.
Update comment in __init__ about 1-indexing.
Add a preceding _ to top and item_count to mark that they aren’t meant
  to be publicly accessed.
Add a __len__ method to show the number of items in the queue.
Remove try/except in insert and always throw a TypeError for mismatched values.
Remove limit printing from the error message; it can be found with size()
  if needed.
Add new nodes to the beginning of the chain at each priority instead of the end.