Update tests per code review.
Remove one float test (testing a non-integer float).
Modify test_create_float so that it fails when creating a queue of size 5.0.
Remove test_find_min_priority_chain because it relies on the order of elements.
Modify test_insert_collision to use a set and not test specific order
Correct a find/replace error of ‘prioritys’ to ‘priorities’ on lines 72 and 96.
Modify test_extract_min_same_priority to use a set and not test specific order.
Correct q.item_count to len(q) on line 87.
Add test_insert_3_with_1_collision, where the non-colliding element is inserted
  second.
Add 5 tests for len()