Fix bug in reconstruct_subset.
The capacity is being detected from the size of the first row of the
max_values table, but the length of the row (the number of columns) is
related to the number of items, not the capacity. Change this to check
the number of rows (the size of the max_values table) so that it will
correctly determine the capacity.