Fix bug in knapsack involving answer table.
The max_values table is only filling the last row, where the capacity is
equal to the knapsack's full capacity. Add a loop over the size of
capacity (from 0 to capacity + 1) so that the table will track the best
partial solutions as well. This will the table to return a better answer
in cases where a greedy, suboptimal answer is an option.