Update factorize and tests per code review.
Factorize factors repeat factors (such as squares) correctly because it
moves through the factors and will not find the second multiple because
it is not a prime. For example, a test for 27 returns the desired result
3 by leaving p (the number that is divided each time) as 9. Since 9 is
not prime, the loop runs until it has tested all primes up to 27 and
then quits.
Change factorize such that it adds a factor to the list multiple times
if it is a repeat factor, such as a square or cube. This requires
changing of the above case so that the remaining 9 gets divided again
down to 3 and 3.