Make changes per code review.
In sieve, include 0 and 1 in the possible_primes list, but set them to
None. This allows for 0 indexing.
Add a test for is_prime to test the number of primes up to 100.
Change is_prime to test some small numbers specifically, check the last
digit of the number for definite composites, and use sieve up to the
square of the number to check for primality.