Begin making changes to SAT per code review.
Add documentation to show that ! is not a valid literal.
Rename some 'unsatisfied' variables and functions for clarity.
Break clause parsing into its own function.
Change assignment[swap] ^= True to assignment[swap] = not
assignment[swap] in two places.
Change check_is_satisfied into a list comprehension.
Change 'negated' to 'positive' in is_satisfied.
Change an if statement in is_satisfied to a < check.
In is_satisfied, change an XOR with negated to an AND with positive.