Fix bug in shortest path involving start vertex.
Start is included in the visited set, but all edges from start are also
added to the list of distances without checking the visited set. This
means that if there is a self loop from start, start will have a known
distance of whatever the loop edge weight is and will not be
re-examined. Change this to check that the edges from start do not lead
back to start before adding them.
Break up some lines in shortest_path and a_star to keep the column
width under 80.