Make some optimizations to cover.
Instead of choosing a random edge, create a function that looks for the
edge that affects the most vertices.
Don't add vertices to the cover if they only have one edge, unless the
edge is the only one connected to either vertex.
Keep track of 'covered' vertices that are directly connected to the
cover by an edge. These vertices don't need to be catered to when
looking for new edges.
Create a function that returns the number of uncovered vertices adjacent
to a vertex.