Fix some bugs in make_hull.
Change the append of the second point (from the start point) so that the
angle is not appended to the point list.
Change the return from a list to a set.
While counting interior points, the while loop checks for turn being <
0. Change this to <= 0 so that we can also overwrite points on the
boundary of the convex hull, which we aren't including in our results.